synchronizer: move db logic into it's own file to declutter main.rs
